Starting working with interface IO
